chore: rename index.ts for systemjs
