+++ src/carousel/carousel.module.ts
