+++ src/index.ts
@@ -1,22 +1,19 @@
 import {NgModule} from '@angular/core';
 
-import {NgbAccordionModule} from './accordion';
-import {NgbAlertModule} from './alert';
-import {NgbButtonsModule} from './buttons';
-import {NgbCarouselModule} from './carousel';
-import {NgbCollapseModule} from './collapse';
-import {NgbDropdownModule} from './dropdown';
-import {NgbPaginationModule} from './pagination';
-import {NgbPopoverModule} from './popover';
-import {NgbProgressbarModule} from './progressbar';
-import {NgbRatingModule} from './rating';
-import {NgbTabsetModule} from './tabset';
-import {NgbTimepickerModule} from './timepicker';
-import {NgbTooltipModule} from './tooltip';
-import {NgbTypeaheadModule} from './typeahead';
-
-export {NgbPanelChangeEvent} from './accordion';
-export {NgbTabChangeEvent} from './tabset';
+import {NgbAccordionModule, NgbPanelChangeEvent} from './accordion/accordion.module';
+import {NgbAlertModule} from './alert/alert.module';
+import {NgbButtonsModule} from './buttons/radio.module';
+import {NgbCarouselModule} from './carousel/carousel.module';
+import {NgbCollapseModule} from './collapse/collapse.module';
+import {NgbDropdownModule} from './dropdown/dropdown.module';
+import {NgbPaginationModule} from './pagination/pagination.module';
+import {NgbPopoverModule} from './popover/popover.module';
+import {NgbProgressbarModule} from './progressbar/progressbar.module';
+import {NgbRatingModule} from './rating/rating.module';
+import {NgbTabsetModule, NgbTabChangeEvent} from './tabset/tabset.module';
+import {NgbTimepickerModule} from './timepicker/timepicker.module';
+import {NgbTooltipModule} from './tooltip/tooltip.module';
+import {NgbTypeaheadModule} from './typeahead/typeahead.module';
 
 @NgModule({
   exports: [
