+++ src/progressbar/progressbar.module.ts
