+++ src/rating/rating.module.ts
