+++ src/tabset/tabset.module.ts
