+++ src/typeahead/typeahead.module.ts
